+++ backend/tests/server/httpserver_get_specification_test.tsx
@@ -1,75 +1,39 @@
-import { expect, it, beforeAll, afterAll } from '@jest/globals'
+import { it, expect, beforeAll, afterAll } from 'vitest'
 import { startModbusTCPserver, stopModbusTCPServer } from '../../src/server/modbusTCPserver.js'
-
 import { HttpErrorsEnum, ImodbusSpecification } from '../../src/shared/specification/index.js'
-import { FakeMqtt, FakeModes, setConfigsDirsBackendTCPForTest, initBussesForTest } from './configsbase.js'
-import supertest from 'supertest'
 import { apiUri } from '../../src/shared/server/index.js'
-import { HttpServer } from '../../src/server/httpserver.js'
-import { Config } from '../../src/server/config.js'
-import { ConfigBus } from '../../src/server/configbus.js'
-import { MqttClient } from 'mqtt'
-import { join } from 'path'
 import { ConfigSpecification } from '../../src/specification/index.js'
-import { MqttSubscriptions } from '../../src/server/mqttsubscriptions.js'
-import { MqttConnector } from '../../src/server/mqttconnector.js'
-import { TempConfigDirHelper } from './testhelper.js'
-let httpServer: HttpServer
-let tempHelper: TempConfigDirHelper
+import { createTestServer, TestServer } from './httpTestHelper.js'
 
-beforeAll(async() => {
+let ts: TestServer
 
-  // fake MQTT: avoid reconnect
-    setConfigsDirsBackendTCPForTest()
-    tempHelper = new TempConfigDirHelper('httpserver_get_spec')
-    tempHelper.setup()
-
-    const conn = new MqttConnector()
-    const msub = new MqttSubscriptions(conn)
-
-    const fake = new FakeMqtt(msub, FakeModes.Poll)
-    conn.getMqttClient = function (onConnectCallback: (connection: MqttClient) => void) {
-      onConnectCallback(fake as any as MqttClient)
-    }
-    new ConfigSpecification().readYaml()
-    const cfg = new Config()
-    await cfg.readYamlAsync()
-      ConfigBus.readBusses()
-      initBussesForTest()
-      HttpServer.prototype.authenticate = (req, res, next) => {
-        next()
-      }
-    await startModbusTCPserver(ConfigSpecification.configDir, ConfigSpecification.dataDir, 0)
-          httpServer = new HttpServer(join(ConfigSpecification.configDir, 'angular'))
-          httpServer.setModbusCacheAvailable()
-          httpServer.init()
+beforeAll(async () => {
+  ts = await createTestServer({ name: 'httpserver_get_spec', backendTCP: true })
+  await startModbusTCPserver(ConfigSpecification.configDir, ConfigSpecification.dataDir, 0)
 })
 afterAll(() => {
   stopModbusTCPServer()
-  if (tempHelper) tempHelper.cleanup()
+  ts.cleanup()
 })
 
 it('Discrete Inputs definition provided check', async () => {
-  if (httpServer){
-    const response = await supertest(httpServer['app'])
-      .get(apiUri.modbusSpecification + '?busid=0&slaveid=3&spec=lc-technology-relay-input')
-      .expect(HttpErrorsEnum.OK)
-    const spec: ImodbusSpecification = response.body  
-        expect(spec.entities).toBeDefined()
-        expect(spec.entities.length).toEqual(16)
-        expect(spec.entities[0].registerType).toEqual(2)
-  }
+  const response = await ts
+    .request()
+    .get(apiUri.modbusSpecification + '?busid=0&slaveid=3&spec=lc-technology-relay-input')
+    .expect(HttpErrorsEnum.OK)
+  const spec: ImodbusSpecification = response.body
+  expect(spec.entities).toBeDefined()
+  expect(spec.entities.length).toEqual(16)
+  expect(spec.entities[0].registerType).toEqual(2)
 })
 
 it('Coils definition provided check', async () => {
-  if (httpServer){
-    const response = await supertest(httpServer['app'])
-      .get(apiUri.modbusSpecification + '?busid=0&slaveid=3&spec=lc-technology-relay-input')
-      .expect(HttpErrorsEnum.OK)
-        const spec: ImodbusSpecification = response.body
-
-    expect(spec.entities).toBeDefined()
-    expect(spec.entities.length).toEqual(16)
-    expect(spec.entities[8].registerType).toEqual(1)
-  }
+  const response = await ts
+    .request()
+    .get(apiUri.modbusSpecification + '?busid=0&slaveid=3&spec=lc-technology-relay-input')
+    .expect(HttpErrorsEnum.OK)
+  const spec: ImodbusSpecification = response.body
+  expect(spec.entities).toBeDefined()
+  expect(spec.entities.length).toEqual(16)
+  expect(spec.entities[8].registerType).toEqual(1)
 })